+++ src/main/qtclient.js
@@ -1,6 +1,8 @@
+var assert = require('assert')
 var bitcoin = require('bitcoin')
 
 function connect (rpcConfig, callback) {
+  assert(rpcConfig, 'Must pass in rpcConfig')
   var client = new bitcoin.Client(rpcConfig)
 
   // test connection