+++ src/main/settings/default.js
@@ -1,0 +1,5 @@
+// these settings need to be changed for environments
+module.exports = {
+  blockCheckInterval: 30 * 1000,
+  mempoolCheckInterval: 10 * 1000
+}