+++ src/main/settings/qtconfig.js
@@ -1,46 +1,37 @@
+var assign = require('object-assign')
 var fs = require('fs-extra')
 var ini = require('ini')
-var path = require('path-extra')
-var _ = require('lodash')
-
-// we use Bitcoin testnet to test capabilities since Bitcoin testnet coins
-// are easy and free to acquire
-var TEST_FILE = path.join(path.datadir('Bitcoin'), 'bitcoin.conf')
-var PROD_FILE = path.join(path.datadir('BlackCoin'), 'blackcoin.conf')
 
 var DEFAULT_SETTINGS = {
   server: 1
 }
 
-function readSync (isTest) {
-  var text = ''
-  try {
-    text = fs.readFileSync(isTest ? TEST_FILE : PROD_FILE).toString('utf8')
-  } catch (e) {}
+function readSync (file) {
+  var text = fs.readFileSync(file).toString('utf8')
+  var data = ini.decode(text)
+  assign(data, DEFAULT_SETTINGS)
 
-  var data = ini.parse(text)
-  data = _.assign(data, DEFAULT_SETTINGS)
-
-  if (isTest) {
-    data.testnet = 1
-  }
+  // rpc info may already be set
+  data.rpcuser = data.rpcuser || 'rpcuser'
+  data.rpcpassword = data.rpcpassword || 'obsidian'
 
   return data
 }
 
-function saveSync (data, isTest) {
-  if (isTest) {
-    data.testnet = 1
-  }
+// if defaults aren't set, this will then save them
+function readSaveSync (file) {
+  var data = readSync(file)
+  saveSync(file, data)
+  return data
+}
 
+function saveSync (file, data) {
   var text = ini.encode(data)
-
-  fs.outputFileSync(isTest ? TEST_FILE : PROD_FILE, text)
+  fs.outputFileSync(file, text)
 }
 
 module.exports = {
-  TEST_FILE: TEST_FILE,
-  PROD_FILE: PROD_FILE,
   readSync: readSync,
+  readSaveSync: readSaveSync,
   saveSync: saveSync
 }